AMASH.Config

Documentation

getIP

Arguments

:: IO String
The IP read from the env var AMASH_MONGO_IP or "127.0.0.1" if the var is empty.

Returns the IP used to connect to the AMASH MongoDB database.

getPort

Arguments

:: IO PortID
The Port read from the env var AMASH_MONGO_PORT or "27017" if the var is empty.

Returns the Port used to connect to the AMASH MongoDB database.

getUser

Arguments

:: IO Text
The User read from the env var AMASH_MONGO_USER or "admin" if the var is empty.

Returns the User used to authenticate at the AMASH MongoDB database.

getPassword

Arguments

:: IO Text
The IP read from the env var AMASH_MONGO_PW or "123" if the var is empty.

Returns the Password used to authenticate at the AMASH MongoDB database.