AMASH.MongoDB.Connection

Documentation

openConnection

Arguments

:: IO Pipe
The pipe used to communicate with the database.

Connect to the database specified in the environment variables. Can fail due to connection errors.

authenticate

Arguments

:: Pipe
The pipe used to communicate with the database.

-> IO Bool
Whether the authentication was successful.

Authenticate at the database using the credentials specified in the environment variables.