AMASH.MongoDB.Helpers

Documentation

thereAreNoOlderEntries

Arguments

:: IO (Bool, Maybe a)
(True, Nothing).

Prints out that there are no earlier rankings to compare the new rankings with. Utility function for "checkIfRankingsAreNew".

getLastSavedDataAndCompare :: Pipe -> Action IO [t] -> (t -> IO (Bool, Maybe a)) -> IO (Bool, Maybe a)

Fetches the last saved data and (if it exists) calls the function that compares it with the new data.

compareFetchedAndOldData :: Val a => a -> Label -> Document -> IO (Bool, Maybe ObjectId)

Compares fetchedData with last saved data. Needs attributeName to extract data from document.

eqMaybe :: Eq a => Maybe a -> a -> Bool

Compares something wrapped in a Maybe with another value and returns false if it doesn't exist.