AMASH.MongoDB.Querys.GetLastSaved

Documentation

getLastSavedRankings

Arguments

:: MonadIO m

=> Text
The application.

-> Action m [Document]
The resulting action.

Builds an action that fetches the last saved rankings entry of an application and category.

getLastSavedVendorEntry

Arguments

:: MonadIO m

=> Text
The vendor id.

-> Text
The collection name.

-> Action m [Document]
The resulting action.

getLastSavedVendorContacts :: MonadIO m => Text -> Action m [Document]

Builds an action that fetches the last saved vendor contacts entry of a vendor id.

getLastSavedVendorMetaData :: MonadIO m => Text -> Action m [Document]

Builds an action that fetches the last saved vendor metadata entry for a vendor id.

getLastSavedAppEntry

Arguments

:: MonadIO m

=> Text
The vendor id.

-> Text
The collection name.

-> Action m [Document]
The resulting action.