AMASH.MongoDB.Querys.Select

Documentation

selectByObjectId :: (Select aQueryOrSelection, Val v) => Collection -> v -> aQueryOrSelection

Selects a document by its objectId from a given collection.

selectApplication

Arguments

:: Select aQueryOrSelection

=> Application
The application.

-> aQueryOrSelection
The resulting selection.

Builds a selection for an application from the rankings database.

selectVendor

Arguments

:: Select aQueryOrSelection

=> Text
The vendor key.

-> aQueryOrSelection
The resulting selection.

Builds a selection for a vendor document from the vendors database.