AMASH.Util

Documentation

showRanking :: Application -> AppsListFilter -> String

rankingsCollectionName :: Show a => Application -> a -> [Char]

elemAtLeastOne

Arguments

:: Eq a

=> [a]
The first list.

-> [a]
The second list against which the elements of the first list are tested.

-> Bool
Whether at least one element of the first list is also an element of the second list.

Given two lists tests if at least one element of the first list is also an element of the second list.