Lib

Description

A Lib module.

Documentation

square

Arguments

:: Num a

=> a
The number

-> a
The square

Calculate the square of a number.

fetchVendorMetaData :: String -> IO ()

Fetches and prints the metadata for a given vendor id (e.g. "1210714")

fetchPluginMetaData :: String -> IO ()

Fetches and prints the metadata for a given plugin key (e.g. "de.scandio.confluence.plugins.pocketquery")